Fix all remaining -Wall -Wextra warnings
Unused variables:
- tmlog_base.c, tmlog_tornbit.c, stats.c: mark block_addr, val, i,
  readindex_checkpoint, statset_all with __attribute__((unused))
- cm.h: mark modedata and w unused (only used under conditional macros)
- local.c: mark buf unused in both functions
- beginend-bits.h: mark trycommit_transaction unused

Format / type fixes:
- tmlog_base.c, tmlog_tornbit.c: add <inttypes.h>; use PRIu64 for uint64_t
- barrier-bits.h: fix %d → %lu for mtm_word_t (unsigned long)
- memset.h: change loop var int i → size_t i (sign-compare vs size_t BUFSIZE)

Missing return statements:
- mode.c: return (mtm_mode_t)-1 at end of mode string lookup
- pwbetl.c: return M_R_SUCCESS from mtm_pwbetl_destroy
- tmlog_base.c, tmlog_tornbit.c: return M_R_SUCCESS from report_stats
- rwlock.c: return 0 from all five stub functions
- heap.cc: return 0 from Heap::init()
- main.c: return NULL from reader/writer thread functions

Incompatible pointer types:
- pwb.c, beginend.c: cast tx->jb to (mtm_jmpbuf_t*) for _ITM_siglongjmp
- memcpy.c, memset.c: add volatile to load/store_bytes forward declarations
- gcc-abi.c: introduce pmalloc_undo_wrapper to avoid cast-function-type

Other:
- tmlog_tornbit.c: fix sign-compare and parentheses in flush-set loop

+++ src/library/mtm/include/mode/common/memset.h
@@ -45,7 +45,7 @@ void _ITM_CALL_CONVENTION _ITM_memset##VARIANT(         void *dst,             \
   mtm_tx_t *tx = mtm_get_tx();						       \
   volatile uint8_t *daddr=dst;                                                 \
   uint8_t          buf[BUFSIZE];                                               \
-  int              i;                                                          \
+  size_t           i;                                                          \
                                                                                \
   if (size == 0) {                                                             \
     return;                                                                    \
